+++ django_afip/templatetags/django_afip.py
@@ -1,19 +1,8 @@
 import re
-import warnings
 
 from django import template
 
 register = template.Library()
-
-
-@register.filter
-def receiptnumber(receipt):
-    warnings.warn(
-        "Use receipt.formatted_number instead.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    return receipt.formatted_number
 
 
 @register.filter